Make generated method calling code more efficient
Previously each function call would: Generate the target, dupe the target object, store it, generate the method, load the target object, generate the paramaters, then call the method.

I have combined a lot of this into a single instruction, so now the flow is: Generate the target, generate the paramaters, call the method using a single invoke dynamic.

I have also changed the representation of messages in the generated classes, so a message add: x to: y will now have the name add$to$ instead of add:to:
The reason for this is that ":" is a special character in dynalink, so you cannot use it within method names when doing dyn:callMethod:x
Another benefit to this is that "$" can be used in Java method names, so I can now write the rt package directly in java

The lexer now automatically changes ":" to "$" in message calls