Make loops actually work
Originally I had planned to use the Java -noverify flag to avoid dealing with stackMaps.
It turns out that it is fairly simple to get ASM to add stackMaps to already generated classes, so now I can run my classes with the verifier, which is better for error detection and is better practice.
I have also modified the while loop generated code, as it lead to error