Fix arrays and DNE bootstrapper
Previously the DNE bootstrapper created invocations with no guard, this was an oversight as it meant that the invocations would stay valid even if the target object changed, leading to the wrong method being called.
I have fixed this by adding a guard to check that the receiver is of exactly the same class (not subclass, as a subclass may have the method trying to be called)

Also, the generated array bytecode had a bug where you couldn't pass anything but a literal number to the "Array new: x" construction. This has now been fixed so that x can be any expression